DAO (Not good) * Fixed wrong name in sql row (artist_id) * Changed ID from int to long.

+++ backend/src/main/java/no/kristiania/munchMoment/database/ArtistRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.ListCrudRepository;
 
 import java.util.List;
 
-public interface ArtistRepository extends ListCrudRepository<Artist, Integer> {
+public interface ArtistRepository extends ListCrudRepository<Artist, Long> {
 
     List<Artist> findArtistsByArtistName(String name);
